hub/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vent-hub"
version = "0.1.0"
description = "Smart vent control hub service"
requires-python = ">=3.11"
dependencies = [
    "aiocoap[all]>=0.4.7",
    "aiohttp>=3.9",
    "cbor2>=5.6",
    "aiosqlite>=0.19",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[project.scripts]
vent-hub = "vent_hub.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vent_hub"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
